backend/requirements.txt
Django
requests
whitenoise
psycopg2-binary
django-environ
djangorestframework
django-cors-headers